backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
elasticsearch==8.10.0
cohere==5.11.3
openai==1.54.3
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart